CasoDeUso-Libreria/backend/requirements.txt
fastapi>=0.110,<1
uvicorn[standard]>=0.23
pydantic>=2.5
python-dotenv>=1
openai>=1.40.0
tenacity>=8.2.3
pydantic
python-dotenv

